requirements.txt
# Core
web3>=6.15.0
asyncpg>=0.29.0
sqlalchemy>=2.0.0
alembic>=1.12.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Graph Analysis
networkx>=3.0

# Machine Learning
scikit-learn>=1.3.0
joblib>=1.3.0

# Async
aiohttp>=3.8.0
asyncio>=3.4.3

# Utilities
redis>=5.0.0
python-json-logger>=2.0.0
tqdm>=4.65.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
pytest-cov>=4.1.0